feat(OnRamp): :recycle: Introduce OnRampRepository.
