+++ lib/domain/repositories/on_ramp.dart
@@ -1,0 +1,10 @@
+import 'package:aewallet/domain/models/onramp.dart';
+
+abstract class OnRampRepository {
+  Future<void> connect();
+  Future<void> dispose();
+  Stream<OnRampEvent> get events;
+  Future<String> get evmAddress;
+
+  Future<num> get maxAmount;
+}
