+++ lib/ui/views/buy/bloc/buy_with_crypto_form_provider.dart
@@ -1,4 +1,5 @@
 import 'package:aewallet/application/onramp/onramp.dart';
+import 'package:aewallet/domain/models/onramp.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
